client/requirements.txt
grpcio
grpcio-tools
httpx

# DEV 
python-dotenv